TileValue variable type changes.

+++ server/src/com/openrsc/server/model/world/region/TileValue.java
@@ -1,13 +1,11 @@
 package com.openrsc.server.model.world.region;
 
 public class TileValue { 
-	
 	public int traversalMask = 0;
-	
-    public int diagWallVal = 0;
-    public int horizontalWallVal = 0;
-    public int overlay = 0;
-    public int verticalWallVal = 0;
-    public int elevation = 0;
+	public int diagWallVal = 0;
+	public byte horizontalWallVal = 0;
+	public byte overlay = 0;
+	public byte verticalWallVal = 0;
+	public int elevation = 0;
 	public boolean projectileAllowed;
 }
